[#129] copy_batch_size config 옵션 적용

+++ src/config.rs
@@ -20,7 +20,7 @@ pub struct Configuraion {
     #[serde(default = "default::peek_changes_timeout_millis")]
     pub peek_changes_timeout_millis: u64,
     #[serde(default = "default::copy_batch_size")]
-    pub copy_batch_size: u64,
+    pub copy_batch_size: usize,
 }
 
 #[derive(Debug, Clone, Deserialize, Serialize)]
@@ -169,8 +169,8 @@ pub mod default {
         SLEEP_MILLIS_AFTER_SYNC_WRITE
     }
 
-    pub const COPY_BATCH_SIZE: u64 = 100000;
-    pub fn copy_batch_size() -> u64 {
+    pub const COPY_BATCH_SIZE: usize = 100000;
+    pub fn copy_batch_size() -> usize {
         COPY_BATCH_SIZE
     }
 }
